# search-manage/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/projects/getDataWithVid/{vid}": {
            "get": {
                "description": "Retrieve project data with a vector id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project/Freelancer"
                ],
                "summary": "Retrieve project data with a vector id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vector Id",
                        "name": "vid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved project data",
                        "schema": {
                            "$ref": "#/definitions/main.ProjectList"
                        }
                    },
                    "400": {
                        "description": "Project ID missing in the path or is not a proper integer",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "description": "Retrieve project data with project id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Project/Freelancer"
                ],
                "summary": "Retrieve project data with project id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved project data",
                        "schema": {
                            "$ref": "#/definitions/main.ProjectList"
                        }
                    },
                    "400": {
                        "description": "Project ID missing in the path or is not a proper integer",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        },
        "/projects/{id}/find/freelancers": {
            "get": {
                "description": "Find freelancers clost to a given project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Find freelancers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found freelancers close to the input project",
                        "schema": {
                            "$ref": "#/definitions/main.SearchFreelancerResponse"
                        }
                    },
                    "400": {
                        "description": "Project ID missing in the path",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/main.VsDefaultResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.Project": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "skills": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.ProjectList": {
            "type": "object",
            "properties": {
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Project"
                    }
                }
            }
        },
        "main.SearchFreelancerResponse": {
            "type": "object",
            "properties": {
                "D": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "numNeighbors": {
                    "type": "integer"
                },
                "numQueryVectors": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "main.VsDefaultResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}